requirements_a.txt
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
numpy==1.23.2
Pillow==9.2.0
densetorch==0.0.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
